requirements.txt
Flask==2.2.2
tensorflow==2.10.0
opencv-python==4.6.0.66
numpy==1.23.3
Pillow==9.2.0
requests==2.28.1